+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `PlotWidget` now has name, classes and disabled parameters.
+- Added `allow_pan_and_zoom` parameter to allow or disable panning and zooming the plot.
 
 ## [0.2.0] - 2025-03-01